Fix HTML resolve-ref test for source vs UI builds
